Switch to remote rendering with a static build
This commit takes the happo-plugin-gatsby library closer to what
happo-plugin-storybook is doing:

1. invoke a static build
2. inject happo-specific javascript that remote workers understand
3. send the entire build to happo.io

This is better for a few different reasons:
- simpler - no longer a need for building yourself
- most likely faster as we're no longer relying on JSDOM to pre-render
pages
- more robust - results will be closer to what the user sees

This commit introduces a few breaking changes:
- no need for the pre-build
- `pageFilter` is replaced by `pages` - an array of urls to
visit/screenshoot
- `publicFolder` is now known as `outputDir` - which is more in line
with the storybook plugin.

A change was pushed to happo workers as part of testing this commit out.
By default, the `history` API in the browser is mocked. This is done to
prevent errors while running because a script navigated elsewhere. In
the gatsby case however, we rely on replaceState and pushState. So I had
to avoid mocking the history object for these types of runs.